Finalisation controller et test por premiere API creation championnat

+++ src/main/java/com/mrp/app/domain/championship/model/RankEntity.java
@@ -4,11 +4,13 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 public class RankEntity {
     private Long rank;
-    private Float points;
+    private BigDecimal points;
     private PersonEntity attendee;
 }
