+++ src/main/java/com/mrp/app/port/mongo/document/RankDocument.java
@@ -5,11 +5,13 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 public class RankDocument {
     private Long rank;
-    private Float points;
+    private BigDecimal points;
     private PersonDocument attendee;
 }
